backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.0.2
